[store] refactor: remove unused method in trait SledOrderedSerde
