+++ fusestore/store/src/meta_service/raft_state_kv.rs
@@ -86,14 +86,6 @@ impl SledOrderedSerde for RaftStateKey {
 
         Err(ErrorCode::MetaStoreDamaged("invalid key IVec"))
     }
-
-    fn order_preserved_serialize(&self, _buf: &mut [u8]) {
-        unimplemented!("no need")
-    }
-
-    fn order_preserved_deserialize(_buf: &[u8]) -> Self {
-        unimplemented!("no need")
-    }
 }
 
 impl SledSerde for RaftStateValue {}
